// lib/types.ts
export interface PolymarketToken {
  token_id: string
  outcome:  string
  price?:   number
  winner?:  boolean
}

export interface PolymarketMarket {
  id:            string
  condition_id:  string
  question:      string
  description?:  string
  category?:     string
  end_date_iso?: string
  active:        boolean
  closed:        boolean
  volume?:       number
  volume24hr?:   number
  liquidity?:    number
  best_bid?:     number
  best_ask?:     number
  last_trade_price?: number
  outcomes?:     string[]
  outcomePrices: string
  clobTokenIds?: string[]
  tokens?:       PolymarketToken[]
  image?:        string
  icon?:         string
  minimum_tick_size?:  number
  minimum_order_size?: number
  neg_risk?:           boolean
  market_slug?:        string
  maker_base_fee?:     number
  taker_base_fee?:     number
  accepting_orders?:   boolean
  archived?:           boolean
  fpmm?:               string
  game_start_time?:    string
  rewards?: {
    max_spread: number
    min_size:   number
    rates:      any | null
  }
  question_id?:             string
  accepting_order_timestamp?: string | null
  enable_order_book?:       boolean
  is_50_50_outcome?:        boolean
  neg_risk_market_id?:      string
  neg_risk_request_id?:     string
  notifications_enabled?:   boolean
  seconds_delay?:           number
  tags?:                    string[]
}